requirements.txt
# ── Веб-сервер ───────────────────────────────────────────────────────────────
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1
aiofiles==25.1.0
python-multipart==0.0.22
jinja2==3.1.6
markupsafe==3.0.3

# ── База данных ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.48
greenlet==3.3.2

# ── Очередь задач ────────────────────────────────────────────────────────────
huey==2.6.0

# ── Pydantic / типы ───────────────────────────────────────────────────────────
pydantic==2.12.5
pydantic-core==2.41.5
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# ── AI: PyTorch ───────────────────────────────────────────────────────────────
torch==2.10.0
torchaudio
torchvision==0.25.0
triton==3.6.0

# ── AI: CUDA (устанавливаются автоматически вместе с torch, но зафиксированы) ─
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90

# ── AI: Whisper ───────────────────────────────────────────────────────────────
openai-whisper
stable-ts
tiktoken
ctranslate2==4.7.1
tokenizers==0.22.2

# ── AI: Разделение вокала ─────────────────────────────────────────────────────
audio-separator==0.41.1
onnxruntime==1.24.3

# ── Аудио-обработка ───────────────────────────────────────────────────────────
librosa==0.11.0
soundfile==0.12.1
pydub==0.25.1
audioread==3.1.0
soxr==1.0.0
samplerate==0.1.0
resampy==0.4.3
julius==0.2.7
av==16.1.0

# ── Числа / ML утилиты ────────────────────────────────────────────────────────
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.8.0
numba==0.64.0
llvmlite==0.46.0
einops==0.8.2
safetensors==0.7.0
diffq==0.2.4
rotary-embedding-torch==0.6.5

# ── Метаданные / теги ─────────────────────────────────────────────────────────
tinytag==2.2.0

# ── Текст (Genius) ────────────────────────────────────────────────────────────
lyricsgenius==3.10.1
beautifulsoup4==4.14.3
soupsieve==2.8.3

# ── HTTP / сеть ───────────────────────────────────────────────────────────────
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
urllib3==2.6.3
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
h11==0.16.0
anyio==4.12.1

# ── HuggingFace ───────────────────────────────────────────────────────────────
huggingface-hub==1.6.0
hf-xet==1.3.2
fsspec==2026.2.0
filelock==3.25.1

# ── Утилиты ───────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.2.28
tqdm==4.67.3
packaging==26.0
click==8.3.1
rich==14.3.3
pygments==2.19.2
six==1.17.0
decorator==5.2.1
lazy-loader==0.5
pooch==1.9.0
platformdirs==4.9.4
mpmath==1.3.0
sympy==1.14.0
networkx==3.6.1
threadpoolctl==3.6.0
joblib==1.5.3
setuptools==82.0.1
cffi==2.0.0
pycparser==3.0
pillow==12.1.1
msgpack==1.1.2

# ── GUI (pywebview + Qt) ──────────────────────────────────────────────────────
pywebview==5.0.5
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0
qtpy
psutil